v7.3.37 - fix an Aurelium related config warning, addresses github issue #90

+++ src/main/java/net/slipcor/treeassist/externals/AureliumSkillsHook.java
@@ -15,7 +15,7 @@ public class AureliumSkillsHook {
      * @param player the player who should gain the EXP
      * @param value  the EXP amount to give
      */
-    public static void addAureliumExp(Player player, Integer value) {
+    public static void addAureliumExp(Player player, Double value) {
         AureliumAPI.addXp(player, Skills.FORAGING, value);
     }
 }
